find5: Remove all remaining card emulation items
The NFCEE access XMLs shouldn't be here since they are
for accessing a secure element we don't have.  Forgot
to remove these with the HCE patch.

Also, upon further investigation, HCE isn't fully implemented
on NXP chips.  All HCE-related functions in the nxp part of
packages/apps/Nfc are stubs that do nothing, so even if
HOST_EMULATION is turned on when building external/libnfc-nxp
it doesn't actually work.  A reader is detected but nothing
else happens when trying to pay.

As a result, remove the HCE permissions so people don't wind
up in an awkward situation where they think they'll pay with
their phone but it does nothing when they tap.

Change-Id: Icb605420559d13e1848f057cf646ced31aa7c8d3

Conflicts:
	device.mk